src/python/requirements.txt
flask==2.3.2
requests==2.31.0
opentelemetry-sdk==1.25.0
opentelemetry-api==1.25.0
opentelemetry-instrumentation==0.46b0